* remove a number of warnings. We are left with mostly "unhandled in switch" warnings   which should either prompt us to complete implementations or add default: were it   is in fact correct
git-svn-id: svn+ssh://c4science.ch/source/amie/trunk@2917 64fb8e13-074e-4207-b257-2d34fb7690c6

+++ examples/main_thermalexpansion.cpp
@@ -124,7 +124,7 @@ Vector thermalDeformation(Vector matrixProp, Vector inclusionProp, int nInc, dou
 	std::vector<Inclusion *> inc = buildInclusion(nInc,pInc) ;
 	
 	sample.setBehaviour(mTDef) ;
-	for(int i = 0 ; i < inc.size() ; i++)
+	for(size_t i = 0 ; i < inc.size() ; i++)
 	{
 		  inc[i]->setBehaviour(iTDef) ;
 		  F.addFeature(&sample, inc[i]) ;
